+++ yorkie/src/main/kotlin/dev/yorkie/document/change/CheckPoint.kt
@@ -8,7 +8,7 @@ import kotlin.math.max
  */
 internal data class CheckPoint(
     val serverSeq: Long,
-    val clientSeq: Int,
+    val clientSeq: Long,
 ) {
 
     /**
